engine: restructuring sstable wrapper
