+++ lib/db/sstable_wrapper.cc
@@ -18,29 +18,46 @@
 
 using namespace dibibase::db;
 
-SSTableWrapper::SSTableWrapper()
-    : m_last_sstable_id(0), m_logger(Logger::make()) {
+SSTableWrapper::SSTableWrapper(std::multimap<std::string, std::string> memtable)
+    : m_memtable(memtable), m_sstable_id(0), m_logger(Logger::make()) {
 
-  // Fetching the last SSTable ID.
+  fetch_recent_sstable_id();
+
+  //Creating new sstable ID.
+  m_sstable_id++;
+
+  char *buffer = new char[4096];
+  size_t allocated_bytes = encode_data(buffer);
+  m_files = std::unique_ptr<SSTableFiles>(
+      new SSTableFiles(buffer, allocated_bytes, m_sstable_id));
+
+  delete[] buffer;
+}
+
+SSTableWrapper::SSTableWrapper() : m_sstable_id(0), m_logger(Logger::make()) {
+  fetch_recent_sstable_id();
+}
+
+void SSTableWrapper::fetch_recent_sstable_id() {
   m_fd_metadata = open("sstables/metadata.db", O_RDONLY);
 
   if (m_fd_metadata > 0) {
-    char *buffer = new char[2];
+    char *mem_buff = new char[2];
 
-    ssize_t read_bytes = read(m_fd_metadata, buffer, 4096);
+    ssize_t read_bytes = read(m_fd_metadata, mem_buff, 2);
     if (read_bytes < 0) {
       m_logger.err("Failed to read metadata.");
     } else {
-      m_last_sstable_id = buffer[0] + (buffer[1] << 8);
+      m_sstable_id = mem_buff[0] + (mem_buff[1] << 8);
     }
 
-    delete[] buffer;
+    delete[] mem_buff;
   }
 }
 
 std::string SSTableWrapper::read_key_value_pair(std::string key) {
   // In case SSTables haven't been created yet.
-  if (m_last_sstable_id == 0) {
+  if (m_sstable_id == 0) {
     m_logger.info("No SSTables found.");
     return "";
   }
@@ -48,8 +65,8 @@ std::string SSTableWrapper::read_key_value_pair(std::string key) {
   std::string required_value;
 
   // Iterating through all SStables starting from the last created one.
-  for (int i = m_last_sstable_id; i > 0; --i) {
-    m_fetched_files = std::unique_ptr<SSTableFiles>(new SSTableFiles(i));
+  for (int i = m_sstable_id; i > 0; --i) {
+    m_files = std::unique_ptr<SSTableFiles>(new SSTableFiles(i));
     std::string value = decode_data(key);
 
     // In case key is found in the current SSTable.
@@ -65,7 +82,7 @@ std::string SSTableWrapper::read_key_value_pair(std::string key) {
 std::string SSTableWrapper::decode_data(std::string required_key) {
 
   char *buffer = new char[4096];
-  ssize_t read_bytes = m_fetched_files->read_data_file(buffer);
+  ssize_t read_bytes = m_files->read_data_file(buffer);
   if (read_bytes <= 0) {
     return "";
   }
@@ -102,6 +119,60 @@ std::string SSTableWrapper::decode_data(std::string required_key) {
 
   delete[] buffer;
   return required_value;
+}
+
+size_t SSTableWrapper::encode_data(char *buffer) {
+  // Data Format:
+  // First 2 bytes represents nu. of entires in memtable.
+  // For each entry:
+  // - 2 bytes for the key length.
+  // - key characters, each takes 1 byte.
+  // - 2 bytes for the value length.
+  // - value characters, each takes 1 byte.
+
+  size_t allocated_bytes = 0;
+
+  uint16_t mem_size = m_memtable.size();
+  char mem_buff[2];
+  memset(mem_buff, 0, sizeof(mem_buff));
+  memcpy(mem_buff, reinterpret_cast<char *>(&mem_size), 2);
+  memcpy(buffer, mem_buff, 2);
+
+  buffer += 2;
+  allocated_bytes += 2;
+
+  for (auto it : m_memtable) {
+    // Storing key size in 2 bytes.
+    uint16_t key_size = it.first.size();
+    char key_buff[2];
+    memset(key_buff, 0, sizeof(key_buff));
+    memcpy(key_buff, reinterpret_cast<char *>(&key_size), 2);
+    memcpy(buffer, key_buff, 2);
+
+    buffer += 2;
+    allocated_bytes += 2;
+
+    // Storing key characters, each character in 1 byte.
+    std::strcpy(buffer, it.first.c_str());
+    buffer += key_size;
+    allocated_bytes += key_size;
+
+    // Storing value size in 2 bytes.
+    uint16_t value_size = it.second.size();
+    char value_buff[2];
+    memset(value_buff, 0, sizeof(value_buff));
+    memcpy(value_buff, reinterpret_cast<char *>(&value_size), 2);
+    memcpy(buffer, value_buff, 2);
+
+    buffer += 2;
+    allocated_bytes += 2;
+
+    // Storing value characters, each character in 1 byte.
+    std::strcpy(buffer, it.second.c_str());
+    buffer += value_size;
+    allocated_bytes += value_size;
+  }
+  return allocated_bytes;
 }
 
 SSTableWrapper::~SSTableWrapper() {
